Fix tabLayout bug.
