+++ app/src/main/java/com/github/monolabhituszi/portal/ui/home/TimeTableItem.kt
@@ -1,0 +1,8 @@
+package com.github.monolabhituszi.portal.ui.home
+
+import androidx.fragment.app.Fragment
+import com.github.monolabhituszi.portal.R
+
+class TimeTableItem : Fragment(R.layout.fragment_timetable) {
+
+}
